Bunch of refactor and feature work
Features:
* Make the device discovery scheme configurable at runtime;
  this enables reuse of one binary on udevd and init systems
* Allow assuming a single static GPU device; this massively
  accelerates device opening on some systems

Refactor:
* introduce the notion of device acquisition, to more clearly
  separate the udev and mknod approaches
* some doc fixes

This is all very highly explorative.

The whole schema should be evolved much, much further, with
even more focus on udevd and possibly io_ring testing for devices
to pop up.

Before any of that takes place, review to what extend this is even
still applicable in a Wayland world - don't beat on a dead X11 horse
if the future doesn't have these problems.